+++ cis-compute-storage/outputs.tf
@@ -31,9 +31,19 @@ output "block_volumes" {
   value       = var.enable_output ? oci_core_volume.these : null
 }
 
+output "custom_block_volume_backup_policies" {
+  description = "The custom Block Volume backup policies"
+  value       = var.enable_output ? oci_core_volume_backup_policy.custom : null
+}
+
 output "file_systems" {
   description = "The file systems"
   value       = var.enable_output ? oci_file_storage_file_system.these : null
+}
+
+output "file_system_quota_rules" {
+  description = "The file system quota rules"
+  value       = var.enable_output ? oci_file_storage_file_system_quota_rule.these : null
 }
 
 output "file_systems_mount_targets" {